requirements.txt
asgiref==3.7.2
autocommand==2.2.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
httpagentparser==1.9.5
idna==3.4
importlib-resources==6.1.1
more-itertools==9.1.0
packaging==23.1
portend==3.2.0
pydantic==1.10.10
pyTelegramBotAPI==4.15.4
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tempora==5.5.0
typing-extensions==4.7.1
tzdata==2023.3
telebot==0.0.5
python-dotenv==1.0.1
urllib3==2.0.3
